added tests

+++ spring-09/src/main/java/ru/otus/spring/dao/BookDao.java
@@ -6,10 +6,10 @@ import java.util.List;
 
 public interface BookDao {
     int count();
-    int insert(String bookName, int authorId, int genreId);
-    Book getById(int id);
-    List<Book> getByAuthorId(int authorId);
-    List<Book> getByGenreId(int genreId);
+    int insert(String bookName, long authorId, long genreId);
+    Book getById(long id);
+    List<Book> getByAuthorId(long authorId);
+    List<Book> getByGenreId(long genreId);
     List<Book> getAll();
     int update(long id, String name, long author_id, long genre_id);
     void deleteById(long id);
